Move functions.py to use new Magnetometer proto
