+++ pysquared/hardware/magnetometer/lis2mdl/factory.py
@@ -1,7 +1,7 @@
 from ....logger import Logger
 from ...decorators import with_retries
 from ...exception import HardwareInitializationError
-from ..magnetometer_factory_protocol import MagnetometerProto
+from ..magnetometer_protocol import MagnetometerProto
 
 try:
     from mocks.circuitpython.adafruit_lis2mdl.lis2mdl import LIS2MDL  # type: ignore
