+++ pysquared/hardware/magnetometer/magnetometer_protocol.py
